backend/requirements.txt
anthropic==0.52.0
firebase-admin==6.6.0
flask==3.1.0
flask-cors==5.0.1
python-dotenv==1.1.0
